infrastructure/testdriver/execute/test_async.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver execute method (async)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/channel.sub.js"></script>

<script>
setup({single_test: true})
onload = async () => {
    let uuid = "{{uuid()}}";
    let url = "https://{{host}}:{{ports[https][0]}}/infrastructure/testdriver/execute/child_slow.html?uuid=" + uuid;
    win = window.open(url);

    let remote = await new TestDriverRemote(uuid);

    let result = await remote.executeScript(async (elemId) => {
        await new Promise((resolve) => setTimeout(resolve, 1000));
        return document.getElementById(elemId).textContent;
    }, ["test"]);
    assert_equals(result.trim(), "PASS");
    done();
}
</script>
